requirements.txt
requests
python-nmap
tabulate
python-dotenv
flask
python-docx
